Added support for 32 bit windows

+++ psl/inc/psl/memory/raw_region.hpp
@@ -5,7 +5,7 @@
 namespace memory {
 class raw_region {
   public:
-	raw_region(uint64_t size);
+	raw_region(size_t size);
 	~raw_region();
 	raw_region(const raw_region& other);
 	raw_region(raw_region&& other);
@@ -13,7 +13,7 @@ class raw_region {
 	raw_region& operator=(raw_region&& other);
 
 	void* data() const noexcept { return m_Base; }
-	uint64_t size() const noexcept { return m_Size; }
+	size_t size() const noexcept { return m_Size; }
 	size_t pageSize() const noexcept { return m_PageSize; }
 
   private:
