+++ psl/inc/psl/memory/region.hpp
@@ -31,8 +31,8 @@ class region {
 	/// segment (being aligned to pages, and starting at a page).
 	region(region& parent,
 		   memory::segment& segment,
-		   uint64_t pageSize,
-		   uint64_t alignment,
+		   size_t pageSize,
+		   size_t alignment,
 		   allocator_base* allocator = new default_allocator());
 
   public:
@@ -41,7 +41,7 @@ class region {
 	/// \param[in] alignment the alignment value of the region.
 	/// \param[in] allocator the allocator that should be used internally.
 	/// \warning \a allocators should not be shared unless the allocator itself supports such a behaviour.
-	region(uint64_t size, uint64_t alignment, allocator_base* allocator = new default_allocator());
+	region(size_t size, size_t alignment, allocator_base* allocator = new default_allocator());
 
 	~region();
 	region(const region& other) = delete;
@@ -107,20 +107,12 @@ class region {
 
 	bool deallocate(segment& segment);
 	bool deallocate(std::optional<segment>& segment);
-	allocator_base* allocator() const {
-		return m_Allocator;
-	};
+	allocator_base* allocator() const { return m_Allocator; };
 	void compact();
 	void decommit_unused();
-	void* data() const {
-		return m_Base;
-	}
-	uint64_t size() const {
-		return m_Size;
-	}
-	size_t alignment() const {
-		return m_Alignment;
-	}
+	void* data() const { return m_Base; }
+	size_t size() const { return m_Size; }
+	size_t alignment() const { return m_Alignment; }
 	memory::range_t range() const {
 		return memory::range_t {(std::uintptr_t)(m_Base), (std::uintptr_t)(m_Base) + m_Size};
 	}
@@ -139,9 +131,9 @@ class region {
 	allocator_base* m_Allocator {nullptr};
 
 #ifdef PLATFORM_WINDOWS
-	std::pair<uint64_t, uint64_t> page_range(const memory::range_t& range);
+	std::pair<size_t, size_t> page_range(const memory::range_t& range);
 	std::vector<state> m_PageState;
 #endif
-	uint64_t m_PageSize {0u};
+	size_t m_PageSize {0u};
 };
 }	 // namespace memory
